requirements.txt

click
portalocker
openai
GitPython

pyinstaller